🔧 Added black, ruff config
Added black formatting and ruff linting to the project.
The CI now includes a lint job and docs build job.

The addition of black and ruff needed a lot of fixes that have been done.
Some fixes were added to the ruff ignore list to be re-enabled in a future PR.

The PR also removes the docs/requirements.txt file in favor of pyproject.toml.
It was wrong anyway as it lacked all the dependencies to run sphinxcontrib-programoutput (the app itself is called).

A VS Code settings/launch config was added as template files to .vscode/*.template.json. Users can copy them to .vscode/{settings,launch}.json and customize them.