✨ Generate event specific reports (#32)
This PR makes it possible to aggregate the runtime of event listeners
for all fired Sphinx events.
Tricky part in pyinstrument is that all event listeners appear under
`EventManager.emit`, so the event name is lost.
The PR monkeypatches Sphinx to use a custom Event listener which injects
an event specific `emit` function, so it can be tracked in pyinstrument.

The PR can be tested with `sphinx-analysis --project needs
--pyinstrument --tree`.
A new JSON report of name `pyinstrument_sphinx_events.json` is created.
It contains all unique combinations of file path, class and function of
all event listeners, so it will pick up all extensions that have
listeners registered.
The dictionary
`sphinx_performance/sphinx_events.py:CUSTOM_FRAMES_BY_EVENT` can be
modified to add more frames of interest (unrelated to events). It is
currently used to track the runtime of `sphinxcontrib-plantuml`.

The generated JSON file can be used for 2 purposes:
- As a support to the pyinstrument HTML report to get a per-event
aggregation
- To run in a CI and check how a PR to Sphinx or an extension affects
performance

The HTML report now contains event specific function names, so event
names can be searched.